requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Werkzeug==3.0.1
PyJWT==2.8.0
bcrypt==4.1.2
openai==1.99.9
resend==0.6.0

